+++ docs/src/examples/components/Icon/Types/IconExample.shorthand.tsx
@@ -1,25 +1,23 @@
 import React from 'react'
 import { Icon } from 'stardust'
 
-const examples = [
-  'chess rock',
-  'ordered list',
-  'unordered list',
-  'book',
-  'circle',
-  'closed captioning',
-  'compress',
-  'expand',
-  'play',
-  'stop',
-  'calendar alternate outline',
-  'coffee',
-  'compass outline',
-  'area chart',
-]
-
 const IconExample = () => (
-  <div>{examples.map(type => <Icon key={type} size="big" name={type} />)}</div>
+  <div>
+    <Icon size="big" name="chess rock" />
+    <Icon size="big" name="ordered list" />
+    <Icon size="big" name="unordered list" />
+    <Icon size="big" name="book" />
+    <Icon size="big" name="circle" />
+    <Icon size="big" name="closed captioning" />
+    <Icon size="big" name="compress" />
+    <Icon size="big" name="expand" />
+    <Icon size="big" name="play" />
+    <Icon size="big" name="stop" />
+    <Icon size="big" name="calendar alternate outline" />
+    <Icon size="big" name="coffee" />
+    <Icon size="big" name="compass outline" />
+    <Icon size="big" name="area chart" />
+  </div>
 )
 
 export default IconExample