+++ docs/src/examples/components/Icon/Variations/IconExampleBordered.shorthand.tsx
@@ -1,0 +1,18 @@
+import React from 'react'
+import { Icon } from 'stardust'
+
+const IconExampleBordered = () => (
+  <div>
+    <Icon name="chess rock" bordered />
+    <Icon name="book" bordered />
+    <Icon name="expand" bordered />
+    <Icon name="play" bordered />
+    <Icon name="stop" bordered />
+    <Icon name="calendar alternate outline" bordered />
+    <Icon name="coffee" bordered />
+    <Icon name="compass outline" bordered />
+    <Icon name="area chart" bordered />
+  </div>
+)
+
+export default IconExampleBordered