+++ docs/src/examples/components/Icon/Variations/IconExampleCircular.shorthand.tsx
@@ -1,0 +1,18 @@
+import React from 'react'
+import { Icon } from 'stardust'
+
+const IconExampleCircular = () => (
+  <div>
+    <Icon name="chess rock" circular />
+    <Icon name="book" circular />
+    <Icon name="expand" circular />
+    <Icon name="play" circular />
+    <Icon name="stop" circular />
+    <Icon name="calendar alternate outline" circular />
+    <Icon name="coffee" circular />
+    <Icon name="compass outline" circular />
+    <Icon name="area chart" circular />
+  </div>
+)
+
+export default IconExampleCircular